refactor(docker): split compose responsibilities
Move Compose CLI setup, loading, deployment, lifecycle, change detection, labels, and job readiness into focused same-package files without changing behavior. Relocate the Swarm wait helper from production code into its sole test caller.
